Change price column type to integer of table lottery

+++ posttest/init.sql
@@ -6,7 +6,7 @@ DROP TABLE IF EXISTS lottery CASCADE;
 CREATE TABLE lottery (
     id SERIAL PRIMARY KEY,
     ticket VARCHAR(6) NOT NULL,
-    price NUMERIC(10, 2) NOT NULL,
+    price INTEGER NOT NULL,
     amount INTEGER NOT NULL
 );
